+++ .gitignore
@@ -27,4 +27,3 @@ dist-ssr
 
 #File 
 DevApp.tsx
-handlerTest.ts
